add cprofile
