+++ pyproject.toml
@@ -34,11 +34,10 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
   "flake8 ~= 6.0.0",
-]
-test = [
   "pytest ~= 7.4.0",
   "pytest-cov ~= 4.1.0",
   "pytest-asyncio ~= 0.21.0",
+  "flameprof ~= 0.4",
 ]
 
 [project.urls]
